requirements.txt
# PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch==0.7.1

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Experiment Tracking
wandb>=0.15.0
dvc[s3]>=3.0.0
optuna>=3.3.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
streamlit>=1.33.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-multipart>=0.0.6
thop>=0.1.1
icrawler>=0.6.9
